+++ app/models/sale_item_lot_allocation.py
@@ -1,7 +1,8 @@
 import uuid
 from datetime import date, datetime
 from typing import TYPE_CHECKING
-from sqlalchemy import Date, Integer, ForeignKey, DateTime, Numeric, Uuid, func
+from decimal import Decimal
+from sqlalchemy import Date, ForeignKey, DateTime, Numeric, Uuid, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from app.core.database import Base
 
@@ -21,7 +22,7 @@ class SaleItemLotAllocation(Base):
     purchase_item_id: Mapped[uuid.UUID] = mapped_column(
         ForeignKey("purchase_items.id", ondelete="RESTRICT"), nullable=False, index=True
     )
-    quantity_allocated: Mapped[int] = mapped_column(Integer, nullable=False)
+    quantity_allocated: Mapped[Decimal] = mapped_column(Numeric(10, 4), nullable=False)
     unit_cost_snapshot: Mapped[float] = mapped_column(Numeric(14, 2), nullable=False)
     lot_purchase_date: Mapped[date] = mapped_column(Date, nullable=False)
